+++ test/compilationTests/gnosis/MarketMakers/MarketMaker.sol
@@ -8,7 +8,7 @@ contract MarketMaker {
     /*
      *  Public functions
      */
-    function calcCost(Market market, uint8 outcomeTokenIndex, uint outcomeTokenCount) public constant returns (uint);
-    function calcProfit(Market market, uint8 outcomeTokenIndex, uint outcomeTokenCount) public constant returns (uint);
-    function calcMarginalPrice(Market market, uint8 outcomeTokenIndex) public constant returns (uint);
+    function calcCost(Market market, uint8 outcomeTokenIndex, uint256 outcomeTokenCount) public constant returns (uint256);
+    function calcProfit(Market market, uint8 outcomeTokenIndex, uint256 outcomeTokenCount) public constant returns (uint256);
+    function calcMarginalPrice(Market market, uint8 outcomeTokenIndex) public constant returns (uint256);
 }
